CARNET-2; Remove dim style workaround for Windows (#331); <cargo>Remove dim style workaround for Windows (#331)</cargo>
See: https://github.com/chalk/chalk/pull/330/#issuecomment-471977551

The issue seems to have been fixed in newer Windows 10 builds. We're not interested in adding a conditional for older Windows versions as the fix severely complicates the codebase, and it also creates problems for consumers as it makes the output unpredictable.